Mina implementada
Cuando esta apuntando al primero, el juego se ralentiza

+++ simpleSdlWrapper/ebentoak.h
@@ -4,8 +4,9 @@
 
 typedef struct S_POSIZIOA
 {
-  int x;
-  int y;
+  float x;
+  float y;
+
 } POSIZIOA;
 
 /*
